requirements (1).txt
streamlit>=1.38.0
openai>=1.35.0
pdfplumber>=0.10.3
python-docx>=1.1.0
python-pptx>=0.6.23
chromadb>=0.5.3
tiktoken>=0.7.0
pillow>=10.0.0
pytesseract>=0.3.10
